[typos] rename files add prefix wiz and remove sufix wizard
