+++ asistente_ladm_col/asistente_ladm_col_plugin.py
@@ -54,34 +54,34 @@ from .utils.decorators import (_db_connection_required,
 
 from .gui.dialogs.dlg_about import AboutDialog
 from .gui.dialogs.dlg_controlled_measurement import ControlledMeasurementDialog
-from .gui.wizards.cadastre.source.create_administrative_source_cadastre_wizard import CreateAdministrativeSourceCadastreWizard
-from .gui.wizards.cadastre.surveying.create_boundaries_cadastre_wizard import CreateBoundariesCadastreWizard
-from .gui.wizards.cadastre.spatial_unit.create_building_cadastre_wizard import CreateBuildingCadastreWizard
-from .gui.wizards.cadastre.spatial_unit.create_building_unit_cadastre_wizard import CreateBuildingUnitCadastreWizard
-from .gui.wizards.cadastre.spatial_unit.create_right_of_way_cadastre_wizard import CreateRightOfWayCadastreWizard
-from .gui.wizards.cadastre.spatial_unit.associate_extaddress_cadastre_wizard import AssociateExtAddressWizard
-from .gui.wizards.cadastre.party.create_col_party_cadastre_wizard import CreateColPartyCadastreWizard
-from .gui.wizards.cadastre.party.create_group_party_cadastre import CreateGroupPartyCadastre
-from .gui.wizards.property_record_card.create_legal_party_prc import CreateLegalPartyPRCWizard
-from .gui.wizards.property_record_card.create_market_research_prc import CreateMarketResearchPRCWizard
-from .gui.wizards.property_record_card.create_natural_party_prc import CreateNaturalPartyPRCWizard
-from .gui.wizards.property_record_card.create_nuclear_family_prc import CreateNuclearFamilyPRCWizard
-from .gui.wizards.cadastre.baunit.create_parcel_cadastre_wizard import CreateParcelCadastreWizard
-from .gui.wizards.cadastre.spatial_unit.create_plot_cadastre_wizard import CreatePlotCadastreWizard
-from .gui.wizards.cadastre.surveying.create_points_cadastre_wizard import CreatePointsCadastreWizard
-from .gui.wizards.property_record_card.create_property_record_card_prc import CreatePropertyRecordCardPRCWizard
-from .gui.wizards.cadastre.rrr.create_responsibility_cadastre_wizard import CreateResponsibilityCadastreWizard
-from .gui.wizards.cadastre.rrr.create_restriction_cadastre_wizard import CreateRestrictionCadastreWizard
-from .gui.wizards.cadastre.rrr.create_right_cadastre_wizard import CreateRightCadastreWizard
-from .gui.wizards.cadastre.source.create_spatial_source_cadastre_wizard import CreateSpatialSourceCadastreWizard
-from .gui.wizards.valuation.create_parcel_valuation_wizard import CreateParcelValuationWizard
-from .gui.wizards.valuation.create_horizontal_property_valuation_wizard import CreateHorizontalPropertyValuationWizard
-from .gui.wizards.valuation.create_common_equipment_valuation_wizard import CreateCommonEquipmentValuationWizard
-from .gui.wizards.valuation.create_building_valuation_wizard import CreateBuildingValuationWizard
-from .gui.wizards.valuation.create_building_unit_valuation_wizard import CreateBuildingUnitValuationWizard
-from .gui.wizards.valuation.create_building_unit_qualification_valuation_wizard import CreateBuildingUnitQualificationValuationWizard
-from .gui.wizards.valuation.create_geoeconomic_zone_valuation_wizard import CreateGeoeconomicZoneValuationWizard
-from .gui.wizards.valuation.create_physical_zone_valuation_wizard import CreatePhysicalZoneValuationWizard
+from .gui.wizards.cadastre.source.wiz_create_administrative_source_cadastre import CreateAdministrativeSourceCadastreWizard
+from .gui.wizards.cadastre.surveying.wiz_create_boundaries_cadastre import CreateBoundariesCadastreWizard
+from .gui.wizards.cadastre.spatial_unit.wiz_create_building_cadastre import CreateBuildingCadastreWizard
+from .gui.wizards.cadastre.spatial_unit.wiz_create_building_unit_cadastre import CreateBuildingUnitCadastreWizard
+from .gui.wizards.cadastre.spatial_unit.wiz_create_right_of_way_cadastre import CreateRightOfWayCadastreWizard
+from .gui.wizards.cadastre.spatial_unit.wiz_associate_extaddress_cadastre import AssociateExtAddressWizard
+from .gui.wizards.cadastre.party.wiz_create_col_party_cadastre import CreateColPartyCadastreWizard
+from .gui.wizards.cadastre.party.dlg_create_group_party_cadastre import CreateGroupPartyCadastre
+from .gui.wizards.property_record_card.wiz_create_legal_party_prc import CreateLegalPartyPRCWizard
+from .gui.wizards.property_record_card.wiz_create_market_research_prc import CreateMarketResearchPRCWizard
+from .gui.wizards.property_record_card.wiz_create_natural_party_prc import CreateNaturalPartyPRCWizard
+from .gui.wizards.property_record_card.wiz_create_nuclear_family_prc import CreateNuclearFamilyPRCWizard
+from .gui.wizards.cadastre.baunit.wiz_create_parcel_cadastre import CreateParcelCadastreWizard
+from .gui.wizards.cadastre.spatial_unit.wiz_create_plot_cadastre import CreatePlotCadastreWizard
+from .gui.wizards.cadastre.surveying.wiz_create_points_cadastre import CreatePointsCadastreWizard
+from .gui.wizards.property_record_card.wiz_create_property_record_card_prc import CreatePropertyRecordCardPRCWizard
+from .gui.wizards.cadastre.rrr.wiz_create_responsibility_cadastre import CreateResponsibilityCadastreWizard
+from .gui.wizards.cadastre.rrr.wiz_create_restriction_cadastre import CreateRestrictionCadastreWizard
+from .gui.wizards.cadastre.rrr.wiz_create_right_cadastre import CreateRightCadastreWizard
+from .gui.wizards.cadastre.source.wiz_create_spatial_source_cadastre import CreateSpatialSourceCadastreWizard
+from .gui.wizards.valuation.wiz_create_parcel_valuation import CreateParcelValuationWizard
+from .gui.wizards.valuation.wiz_create_horizontal_property_valuation import CreateHorizontalPropertyValuationWizard
+from .gui.wizards.valuation.wiz_create_common_equipment_valuation import CreateCommonEquipmentValuationWizard
+from .gui.wizards.valuation.wiz_create_building_valuation import CreateBuildingValuationWizard
+from .gui.wizards.valuation.wiz_create_building_unit_valuation import CreateBuildingUnitValuationWizard
+from .gui.wizards.valuation.wiz_create_building_unit_qualification_valuation import CreateBuildingUnitQualificationValuationWizard
+from .gui.wizards.valuation.wiz_create_geoeconomic_zone_valuation import CreateGeoeconomicZoneValuationWizard
+from .gui.wizards.valuation.wiz_create_physical_zone_valuation import CreatePhysicalZoneValuationWizard
 from .gui.dialogs.dlg_load_layers import DialogLoadLayers
 from .gui.dialogs.dlg_quality import DialogQuality
 from .gui.dialogs.dlg_import_from_excel import DialogImportFromExcel
