+++ asistente_ladm_col/gui/wizards/property_record_card/wiz_create_natural_party_prc.py
